pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "gdd-distribucion"
version = "1.0.0"
description = "Sistema modular de Gestión de Distribución con plugins independientes"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Tu Nombre", email = "correo@example.com"},
]
maintainers = [
    {name = "Tu Nombre", email = "correo@example.com"},
]
keywords = ["distribucion", "gestion", "xsales", "reportes", "plugins", "modular"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Telecommunications Industry",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: Spanish",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.0.0",
    "questionary>=1.10.0",
    "PyYAML>=6.0.0",
    "cryptography>=44.0.0",
    "pycryptodome>=3.20.0",
    "requests-html>=0.10.0",
    "selenium>=4.0.0",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
]

docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.2.0",
]

[project.urls]
Homepage = "https://github.com/tu-org/gdd"
Documentation = "https://github.com/tu-org/gdd#readme"
Repository = "https://github.com/tu-org/gdd.git"
Issues = "https://github.com/tu-org/gdd/issues"

[project.scripts]
gdd = "main:main"
gdd-validate = "validate_deployment:main"

[tool.setuptools]
packages = ["core", "plugins"]

[tool.setuptools.package-data]
plugins = ["**/config.yml", "**/config.yaml", "**/config.json"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directorios
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --strict-markers --tb=short"
markers = [
    "unit: tests unitarios",
    "integration: tests de integración",
    "slow: tests lentos",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.coverage.run]
source = ["core", "plugins"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
